Extract ttheme as a ColorSchemeProvider
